[OrderBook] Move book to order_book_manager
