+++ setup.py
@@ -66,7 +66,6 @@ packages_list = ["octobot_trading.util.initializable",
                  "octobot_trading.producers.simulator.prices_updater_simulator",
                  "octobot_trading.producers.simulator.recent_trade_updater_simulator",
                  "octobot_trading.producers.simulator.ticker_updater_simulator",
-                 "octobot_trading.data.book",
                  "octobot_trading.data.margin_portfolio",
                  "octobot_trading.data.order",
                  "octobot_trading.data.position",
